update files `phenoscore` module
